requirements.txt
streamlit
PyGithub
iointel